refactor: rename controller folder in service folder
