+++ services/sync.go
@@ -1,4 +1,4 @@
-package controllers
+package services
 
 import (
 	"fmt"
